+++ CHANGELOG.md
@@ -17,7 +17,7 @@ in the detailed section referring to by linking pull requests or issues.
 
 #### Added
 
-*
+* Event Framework for Asset entity (#1453)
 
 #### Changed